Update simulation.py
Changed Simulation Class:
__init__ - made the initial self._data an array of empty arrays to allow continuous appending rather than having to overwrite the first value
load_data() - In the master branch, rows 15 and 16 were skipped and row 1 was duplicated. I moved the append out of the if statements so that the next value would always be appended and there would be no duplication
load_data() - Changed the final line to make self._ft a numpy array so it could be easily changed when plotting (in case you wanted to show ms or kHz or do some other transformation).